+++ tumblr-importer.php
@@ -10,16 +10,30 @@ License: GPL v2 - http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
 Text Domain: tumblr-importer
 Domain Path: /languages
 */
+defined( 'ABSPATH' ) || exit;
 
 if ( ! defined( 'WP_LOAD_IMPORTERS' ) && ! defined( 'DOING_CRON' ) ) {
 	return;
 }
 
-require_once ABSPATH . 'wp-admin/includes/import.php';
-require_once ABSPATH . 'wp-admin/includes/admin.php';
+define( 'TUMBLR_IMPORTER_PATH', plugin_dir_path( __FILE__ ) );
 
-require_once __DIR__ . '/class-wp-importer-cron.php';
-require_once __DIR__ . '/class-tumblr-import.php';
+// Load the autoloader.
+if ( ! is_file( TUMBLR_IMPORTER_PATH . '/vendor/autoload.php' ) ) {
+    add_action(
+        'admin_notices',
+        static function () {
+            $message      = __( 'It seems like <strong>Tumblr Importer</strong> is corrupted. Please reinstall!', 'tumblr-importer' );
+            $html_message = wp_sprintf( '<div class="error notice tumblr-importer-error">%s</div>', wpautop( $message ) );
+            echo wp_kses_post( $html_message );
+        }
+    );
+    return;
+}
+require_once TUMBLR_IMPORTER_PATH . '/vendor/autoload.php';
+
+//require_once ABSPATH . 'wp-admin/includes/import.php';
+//require_once ABSPATH . 'wp-admin/includes/admin.php';
 
 /**
  * Tumblr Importer Initialisation routines
@@ -31,7 +45,7 @@ function tumblr_importer_init() {
 	global $tumblr_import;
 	load_plugin_textdomain( 'tumblr-importer', false, dirname( plugin_basename( __FILE__ ) ) . '/languages' );
 
-	$tumblr_import = new Tumblr_Import();
+	$tumblr_import = new \Automattic\TumblrImporter\Tumblr_Import();
 	register_importer( 'tumblr', __( 'Tumblr', 'tumblr-importer' ), __( 'Import posts from a Tumblr blog.', 'tumblr-importer' ), array( $tumblr_import, 'start' ) );
 	if ( ! defined( 'TUMBLR_MAX_IMPORT' ) ) {
 		define( 'TUMBLR_MAX_IMPORT', 20 );
